mrc-remote-instance-weight-transport(rwt-skeleton-move,mechanical_provable): Move init_engine onto RemoteInstanceWeightTransport (cut+paste)
