+++ python/sglang/srt/model_executor/model_runner_components/remote_instance_weight_transport.py
@@ -6,7 +6,9 @@ from typing import Any, Callable, Optional
 
 import torch
 
+from sglang.srt.environ import envs
 from sglang.srt.server_args import ServerArgs
+from sglang.srt.utils.network import NetworkAddress, get_local_ip_auto
 
 logger = logging.getLogger(__name__)
 
@@ -25,3 +27,23 @@ class RemoteInstanceWeightTransport:
     @property
     def model(self) -> torch.nn.Module:
         return self.get_model()
+
+    def init_engine(self):
+        try:
+            from mooncake.engine import TransferEngine
+        except ImportError:
+            logger.warning(
+                "Please install mooncake for using remote instance transfer engine: pip install mooncake-transfer-engine"
+            )
+            return
+        self.engine = TransferEngine()
+        local_ip = get_local_ip_auto()
+        self.engine.initialize(
+            local_ip,
+            "P2PHANDSHAKE",
+            envs.MOONCAKE_PROTOCOL.get(),
+            envs.MOONCAKE_DEVICE.get(),
+        )
+        self.session_id = NetworkAddress(
+            local_ip, self.engine.get_rpc_port()
+        ).to_host_port_str()
